+++ tool/generate-error-constants.js
@@ -107,13 +107,15 @@ function appendDatabseErrorCodes(srcDir, codes) {
 function appendSqlErrorCodes(srcDir, codes) {
   var errorFile = path.join(srcDir, 'sql', 'share', 'errmsg-utf8.txt');
   var contents  = fs.readFileSync(errorFile, 'utf-8');
-  var offset    = Number(contents.match(/start-error-number (\d+)/)[1]);
-  var names     = contents.match(/^([A-Z0-9_]+)/mg).map(fixupCode);
-  var num       = 0;
+  var sections  = contents.split(/^start-error-number (\d+)$/m);
 
-  for (var i = 0; i < names.length; i++) {
-    num = offset + i;
-    codes[num] = names[i];
+  for (var i = 1; i < sections.length; i += 2) {
+    var offset = Number(sections[i]);
+    var names  = sections[i + 1].match(/^([A-Z0-9_]+)/mg).map(fixupCode);
+
+    for (var j = 0; j < names.length; j++) {
+      codes[offset + j] = names[j];
+    }
   }
 
   return codes;